8.03 changes - add transparent water

+++ Win/lodepng.cpp
@@ -727,7 +727,7 @@ static unsigned HuffmanTree_makeTable(HuffmanTree* tree) {
     unsigned l = maxlens[i];
     if(l <= FIRSTBITS) continue;
     tree->table_len[i] = l;
-    tree->table_value[i] = pointer;
+    tree->table_value[i] = (unsigned short)pointer;
     pointer += (1u << (l - FIRSTBITS));
   }
   lodepng_free(maxlens);
@@ -751,7 +751,7 @@ static unsigned HuffmanTree_makeTable(HuffmanTree* tree) {
         unsigned index = reverse | (j << l);
         if(tree->table_len[index] != 16) return 55; /*invalid tree: long symbol shares prefix with short symbol*/
         tree->table_len[index] = l;
-        tree->table_value[index] = i;
+        tree->table_value[index] = (unsigned short)i;
       }
     } else {
       /*long symbol, shares prefix with other long symbols in first lookup table, needs second lookup*/
@@ -768,7 +768,7 @@ static unsigned HuffmanTree_makeTable(HuffmanTree* tree) {
         unsigned reverse2 = reverse >> FIRSTBITS; /* l - FIRSTBITS bits */
         unsigned index2 = start + (reverse2 | (j << (l - FIRSTBITS)));
         tree->table_len[index2] = l;
-        tree->table_value[index2] = i;
+        tree->table_value[index2] = (unsigned short)i;
       }
     }
   }
@@ -3725,7 +3725,7 @@ unsigned lodepng_compute_color_stats(LodePNGColorStats* stats,
   if(!numcolors_done) {
     for(i = 0; i < stats->numcolors; i++) {
       const unsigned char* color = &stats->palette[i * 4];
-      error = color_tree_add(&tree, color[0], color[1], color[2], color[3], i);
+      error = color_tree_add(&tree, color[0], color[1], color[2], color[3], (unsigned int)i);
       if(error) goto cleanup;
     }
   }
@@ -4695,7 +4695,7 @@ static unsigned readChunk_iCCP(LodePNGInfo* info, const LodePNGDecoderSettings* 
                           length, &zlibsettings);
   /*error: ICC profile larger than  decoder->max_icc_size*/
   if(error && size > zlibsettings.max_output_size) error = 113;
-  info->iccp_profile_size = size;
+  info->iccp_profile_size = (unsigned int)size;
   if(!error && !info->iccp_profile_size) error = 100; /*invalid ICC profile size*/
   return error;
 }
@@ -4917,18 +4917,18 @@ static void decodeGeneric(unsigned char** out, unsigned* w, unsigned* h,
     If the decompressed size does not match the prediction, the image must be corrupt.*/
     if(state->info_png.interlace_method == 0) {
       size_t bpp = lodepng_get_bpp(&state->info_png.color);
-      expected_size = lodepng_get_raw_size_idat(*w, *h, bpp);
+      expected_size = lodepng_get_raw_size_idat(*w, *h, (unsigned int)bpp);
     } else {
       size_t bpp = lodepng_get_bpp(&state->info_png.color);
       /*Adam-7 interlaced: expected size is the sum of the 7 sub-images sizes*/
       expected_size = 0;
-      expected_size += lodepng_get_raw_size_idat((*w + 7) >> 3, (*h + 7) >> 3, bpp);
-      if(*w > 4) expected_size += lodepng_get_raw_size_idat((*w + 3) >> 3, (*h + 7) >> 3, bpp);
-      expected_size += lodepng_get_raw_size_idat((*w + 3) >> 2, (*h + 3) >> 3, bpp);
-      if(*w > 2) expected_size += lodepng_get_raw_size_idat((*w + 1) >> 2, (*h + 3) >> 2, bpp);
-      expected_size += lodepng_get_raw_size_idat((*w + 1) >> 1, (*h + 1) >> 2, bpp);
-      if(*w > 1) expected_size += lodepng_get_raw_size_idat((*w + 0) >> 1, (*h + 1) >> 1, bpp);
-      expected_size += lodepng_get_raw_size_idat((*w + 0), (*h + 0) >> 1, bpp);
+      expected_size += lodepng_get_raw_size_idat((*w + 7) >> 3, (*h + 7) >> 3, (unsigned int)bpp);
+      if(*w > 4) expected_size += lodepng_get_raw_size_idat((*w + 3) >> 3, (*h + 7) >> 3, (unsigned int)bpp);
+      expected_size += lodepng_get_raw_size_idat((*w + 3) >> 2, (*h + 3) >> 3, (unsigned int)bpp);
+      if(*w > 2) expected_size += lodepng_get_raw_size_idat((*w + 1) >> 2, (*h + 3) >> 2, (unsigned int)bpp);
+      expected_size += lodepng_get_raw_size_idat((*w + 1) >> 1, (*h + 1) >> 2, (unsigned int)bpp);
+      if(*w > 1) expected_size += lodepng_get_raw_size_idat((*w + 0) >> 1, (*h + 1) >> 1, (unsigned int)bpp);
+      expected_size += lodepng_get_raw_size_idat((*w + 0), (*h + 0) >> 1, (unsigned int)bpp);
     }
 
     state->error = zlib_decompress(&scanlines, &scanlines_size, expected_size, idat, idatsize, &state->decoder.zlibsettings);
@@ -5119,7 +5119,7 @@ static unsigned addChunk_PLTE(ucvector* out, const LodePNGColorMode* info) {
   unsigned char* chunk;
   size_t i, j = 8;
 
-  CERROR_TRY_RETURN(lodepng_chunk_init(&chunk, out, info->palettesize * 3, "PLTE"));
+  CERROR_TRY_RETURN(lodepng_chunk_init(&chunk, out, (unsigned int)info->palettesize * 3, "PLTE"));
 
   for(i = 0; i != info->palettesize; ++i) {
     /*add all channels except alpha channel*/
@@ -5143,7 +5143,7 @@ static unsigned addChunk_tRNS(ucvector* out, const LodePNGColorMode* info) {
       --amount;
     }
     if(amount) {
-      CERROR_TRY_RETURN(lodepng_chunk_init(&chunk, out, amount, "tRNS"));
+      CERROR_TRY_RETURN(lodepng_chunk_init(&chunk, out, (unsigned int)amount, "tRNS"));
       /*add the alpha channel values from the palette*/
       for(i = 0; i != amount; ++i) chunk[8 + i] = info->palette[4 * i + 3];
     }
@@ -5177,7 +5177,7 @@ static unsigned addChunk_IDAT(ucvector* out, const unsigned char* data, size_t d
 
   error = zlib_compress(&zlib, &zlibsize, data, datasize, zlibsettings);
   if(!error) {
-    error = lodepng_chunk_createv(out, zlibsize, "IDAT", zlib);
+    error = lodepng_chunk_createv(out, (unsigned int)zlibsize, "IDAT", zlib);
   }
   lodepng_free(zlib);
   return error;
@@ -5194,7 +5194,7 @@ static unsigned addChunk_tEXt(ucvector* out, const char* keyword, const char* te
   size_t keysize = lodepng_strlen(keyword), textsize = lodepng_strlen(textstring);
   size_t size = keysize + 1 + textsize;
   if(keysize < 1 || keysize > 79) return 89; /*error: invalid keyword size*/
-  CERROR_TRY_RETURN(lodepng_chunk_init(&chunk, out, size, "tEXt"));
+  CERROR_TRY_RETURN(lodepng_chunk_init(&chunk, out, (unsigned int)size, "tEXt"));
   lodepng_memcpy(chunk + 8, keyword, keysize);
   chunk[8 + keysize] = 0; /*null termination char*/
   lodepng_memcpy(chunk + 9 + keysize, textstring, textsize);
@@ -5216,7 +5216,7 @@ static unsigned addChunk_zTXt(ucvector* out, const char* keyword, const char* te
                         (const unsigned char*)textstring, textsize, zlibsettings);
   if(!error) {
     size_t size = keysize + 2 + compressedsize;
-    error = lodepng_chunk_init(&chunk, out, size, "zTXt");
+    error = lodepng_chunk_init(&chunk, out, (unsigned int)size, "zTXt");
   }
   if(!error) {
     lodepng_memcpy(chunk + 8, keyword, keysize);
@@ -5247,7 +5247,7 @@ static unsigned addChunk_iTXt(ucvector* out, unsigned compress, const char* keyw
   }
   if(!error) {
     size_t size = keysize + 3 + langsize + 1 + transsize + 1 + (compress ? compressedsize : textsize);
-    error = lodepng_chunk_init(&chunk, out, size, "iTXt");
+    error = lodepng_chunk_init(&chunk, out, (unsigned int)size, "iTXt");
   }
   if(!error) {
     size_t pos = 8;
@@ -5360,7 +5360,7 @@ static unsigned addChunk_iCCP(ucvector* out, const LodePNGInfo* info, LodePNGCom
                         info->iccp_profile, info->iccp_profile_size, zlibsettings);
   if(!error) {
     size_t size = keysize + 2 + compressedsize;
-    error = lodepng_chunk_init(&chunk, out, size, "iCCP");
+    error = lodepng_chunk_init(&chunk, out, (unsigned int)size, "iCCP");
   }
   if(!error) {
     lodepng_memcpy(chunk + 8, info->iccp_name, keysize);
